Button rewrite in TS.
