--- src/Button/index.d.ts
@@ -1,4 +1,0 @@
-import Button from './Button';
-import StatefulButton from './StatefulButton';
-export default Button;
-export { StatefulButton };
